Rename to Tools
My current understanding of the naming convention is that plugins alter
behaviour in import, and tools export keywords.

after_subtest doesn't alter behaviour, and is an exported keyword even
though its not one directly related to testing.
I took Test2::Tools::Explain as confirmation that this module is a Tool.
Test2::Tools::Explain exports the 'explain' keyword, which is also not
directly related to testing.